requirements.txt
requests==2.21.0
python-dotenv==0.9.1